pyproject.toml
[project]
name = "nanovlm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.2.0",
    "einops>=0.8.1",
    "hf-transfer>=0.1.9",
    "huggingface-hub>=0.35.3",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.1",
    "wandb>=0.22.2",
]
